add overflows-of-long and changed int to long

+++ chemodurov.artem/S1/infix-expression.hpp
@@ -6,7 +6,7 @@ namespace chemodurov
 {
   union OperationsOperandsBraces
   {
-    int operand;
+    long operand;
     Operation operation;
     Brace brace;
   };
